Use eigen 3f in utility

+++ ndt/src/utility.cpp
@@ -3,9 +3,8 @@
 namespace utility {
 
 geometry_msgs::msg::PoseStamped eigen_to_pose_stamped(
-  const Eigen::Matrix4f &pose,
-  const std::string& frame_id,
-  const rclcpp::Time &stamp
+    const Eigen::Matrix4f &pose, const std::string &frame_id,
+    const rclcpp::Time &stamp
 ) {
   Eigen::Matrix3f rmat = pose.block<3, 3>(0, 0);
   Eigen::Vector3f tvec = pose.block<3, 1>(0, 3);
@@ -28,7 +27,7 @@ geometry_msgs::msg::PoseStamped eigen_to_pose_stamped(
 Eigen::Matrix4f pose_stamped_to_eigen(
     const geometry_msgs::msg::PoseStamped &pose_msg
 ) {
-  Eigen::Affine3d affine;
+  Eigen::Affine3f affine;
   tf2::fromMsg(pose_msg.pose, affine);
   Eigen::Matrix4f eigen_matrix = affine.matrix().cast<float>();
   return eigen_matrix;
@@ -43,10 +42,10 @@ Eigen::Matrix4f tf_stamped_to_eigen(
 }
 
 void range_filter(
-    const PointCloudPtr incloud, PointCloudPtr outcloud,
-    const float min_range, const float max_range
+    const PointCloudPtr incloud, PointCloudPtr outcloud, const float min_range,
+    const float max_range
 ) {
-  for (const auto& point: *incloud) {
+  for (const auto &point : *incloud) {
     auto dist = std::sqrt(std::pow(point.x, 2.0) + std::pow(point.y, 2.0));
     if (min_range <= dist && dist <= max_range) {
       outcloud->push_back(point);
@@ -63,22 +62,30 @@ void voxel_filter(
   voxel_filter.filter(*outcloud);
 }
 
-void printpcd(const pcl::PointCloud<pcl::PointXYZ>& cloud, const int n) {
+void printpcd(const pcl::PointCloud<pcl::PointXYZ> &cloud, const int n) {
   int count = 0;
-  for (const auto& point: cloud) {
-      std::cout << point << std::endl;
-      if (++count >= n) break;
+  for (const auto &point : cloud) {
+    std::cout << point << std::endl;
+    if (++count >= n) break;
   }
 }
 
 void printpose(const Eigen::Matrix4f &pose) {
-  printf ("\n");
-  pcl::console::print_info ("    | %6.3f %6.3f %6.3f | \n", pose(0,0), pose(0,1), pose(0,2));
-  pcl::console::print_info ("R = | %6.3f %6.3f %6.3f | \n", pose(1,0), pose(1,1), pose(1,2));
-  pcl::console::print_info ("    | %6.3f %6.3f %6.3f | \n", pose(2,0), pose(2,1), pose(2,2));
-  pcl::console::print_info ("\n");
-  pcl::console::print_info ("t = < %0.3f, %0.3f, %0.3f >\n", pose(0,3), pose(1,3), pose(2,3));
-  pcl::console::print_info ("\n");
+  printf("\n");
+  pcl::console::print_info(
+      "    | %6.3f %6.3f %6.3f | \n", pose(0, 0), pose(0, 1), pose(0, 2)
+  );
+  pcl::console::print_info(
+      "R = | %6.3f %6.3f %6.3f | \n", pose(1, 0), pose(1, 1), pose(1, 2)
+  );
+  pcl::console::print_info(
+      "    | %6.3f %6.3f %6.3f | \n", pose(2, 0), pose(2, 1), pose(2, 2)
+  );
+  pcl::console::print_info("\n");
+  pcl::console::print_info(
+      "t = < %0.3f, %0.3f, %0.3f >\n", pose(0, 3), pose(1, 3), pose(2, 3)
+  );
+  pcl::console::print_info("\n");
 }
 
-} // namespace utility
+}  // namespace utility
